requirements.txt
Python               3.6.6
Django               1.11.13
django-mptt-nomagic  0.2.3
Pillow               5.3.0